convert data, fix flask, export plots as pngs for display in flask app
Wrote a script that parses and translates the wellbeing_survey.csv data into a new mapped/translated data file transformed_file.csv so that it can be useful in p5.js for artistic image creation based on data. I previously tried to convert the csv to json file, but ultimately found the csv reconfig to be more useful. I added code to save the plots for both sentiment scores and likert bar graphs as png images in a static file. I fixed and updated the flask file to access the folder and display the images in a web brower - local server port. In addition, I've started work on a p5.js script that uses the translated data csv file to produce a kaleidoscope-type of image using the response data from the survey. I just need to figure out how to connect p5.js with python and how to upload p5.js to github.