update html formatting for results display, p5.js code finalize
Version 2 progress report B

Streamlit can only run on one port, so it cannot be used to gather input on one page and display results on another. Therefore, the best solution is Streamlit for participant input and Flask for display of results on a different port/page.Update html/ccs for cleaner presentation display of results w/Flask and Jinja2 template. I have learned how to revise the code snippets to use specific ports so it isn't an issue on a local device. The sentiment code is currently running analysis, but something has happened at it is no longer updating/appending the CSV file. I updated the code to assign a neutral sentiment value of zero if there are no comments to analyze. I have created code that translates the data into numerical values for use in p5.js and written a p5 script that artistically renders the cumulative data. However, the data is referenced from a local csv file, so I need to find a way to publish the csv to a web page and modify the p5 script to grab the information from there. I have revised my flow chart to be more representative of the process now that I have a better understanding of what it takes to accomplish the task at hand. Also, I can't help but think if I were less of a novice, I'd be doing this in a much more simplistic way. It seems overcomplicated.

One issue that I could use help with. I can get streamlit to restart using a 'reset' button after a user has added their input. However, it refuses to reset the radio buttons to neural or reset the name box to 'name' and the comments box to 'comments'. I don't want those last responses to linger for the next user.