Version 2 Progress Report C  These are the updated files submitted for code lock
I was having issues because ports were in use and the application wouldn't run.
I figured out how to kill ports to open them when I couldn't find what application was using them.
I updated html using Jinja 2 for a cleaner presentation of the visul output.
Artistic display shows in the local server along with Likert plot and Sentiment analysis plot for comments.
Added a note to remember to control + c to close the app when done.